Conversion from frames to seconds enforced

+++ src/main/java/org/epi/model/Disease.java
@@ -5,7 +5,6 @@ import org.epi.util.Error;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleDoubleProperty;
-import javafx.beans.property.SimpleIntegerProperty;
 
 /** Wrapper class for the disease parameters.*/
 public class Disease {
@@ -16,26 +15,26 @@ public class Disease {
     /** The probability of an infected person dying from the disease.*/
     private final DoubleProperty fatalityRate;
 
-    /** The total duration of the disease in number of frames.*/
-    private final IntegerProperty totalDuration;
+    /** The total duration of the disease in number of seconds.*/
+    private final DoubleProperty totalDuration;
 
     /**
      * The constructor for a disease.
      *
      * @param transmissionRisk The probability of a transmission occurring
      * @param fatalityRate The probability of an infected person dying from the disease
-     * @param totalDuration The total duration of the disease in number of frames
+     * @param totalDuration The total duration of the disease in number of seconds
      * @throws IllegalArgumentException if the given transmissionRisk or fatalityRate is less than {@value Error#MIN_PROB} or more than
      *                                  {@value Error#MAX_PROB} and if the totalDuration is negative
      */
-    public Disease(double transmissionRisk, double fatalityRate, int totalDuration) {
+    public Disease(double transmissionRisk, double fatalityRate, double totalDuration) {
         Error.probabilityCheck(transmissionRisk);
         Error.probabilityCheck(fatalityRate);
         Error.nonNegativeCheck(totalDuration);
 
         this.transmissionRisk = new SimpleDoubleProperty(transmissionRisk);
         this.fatalityRate = new SimpleDoubleProperty(fatalityRate);
-        this.totalDuration = new SimpleIntegerProperty(totalDuration);
+        this.totalDuration = new SimpleDoubleProperty(totalDuration);
     }
 
     //---------------------------- Getters & Setters ----------------------------
@@ -107,7 +106,7 @@ public class Disease {
      *
      * @return {@link Disease#totalDuration}
      */
-    public int getTotalDuration() {
+    public double getTotalDuration() {
         return totalDuration.get();
     }
 
@@ -116,17 +115,17 @@ public class Disease {
      *
      * @return {@link Disease#totalDuration}
      */
-    public IntegerProperty totalDurationProperty() {
+    public DoubleProperty totalDurationProperty() {
         return totalDuration;
     }
 
     /**
      * Setter for {@link Disease#totalDuration}
      *
-     * @param totalDuration the total duration of the disease in number of frames
+     * @param totalDuration the total duration of the disease in number of seconds
      * @throws IllegalArgumentException if the given totalDuration is negative
      */
-    public void setTotalDuration(int totalDuration) {
+    public void setTotalDuration(double totalDuration) {
         Error.nonNegativeCheck(totalDuration);
 
         this.totalDuration.set(totalDuration);
